pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["maksim <knyazev.maksim2852@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
alembic = "^1.13.1"
pydantic-settings = "^2.3.3"
bcrypt = "4.0.1"
asyncpg = "^0.29.0"
pip = "^24.1"
install = "^1.3.5"
python-jose = "^3.3.0"
passlib = "^1.7.4"
pyright = "^1.1.368"
celery = "^5.4.0"
fastapi-cache = {extras = ["redis"], version = "^0.1.0"}
fastapi-cache2 = "^0.2.1"
redis = "^5.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
